fix(plugin): move applyBootstrapContent out of entry point to restore plugin load
PR #352's per-load registration refactor added `export const
applyBootstrapContent` to src/index.ts so unit tests could import it
directly. OpenCode's plugin loader treats every named export from a
plugin entry as an additional plugin factory and invokes each one with
no arguments — `applyBootstrapContent` crashed on `output.system.length`
and the loader aborted the whole plugin, including the intended `default`
export. Users on v2.12.1 saw zero Systematic skills or slash commands;
reverting to v2.12.0 restored everything.

Same class as the v2.5.0 / PR #309 regression where
`INTERNAL_AGENT_SIGNATURES` was exported from the entry. That fix moved
the constant to src/lib/bootstrap.ts with a comment "must NOT be
re-exported from the plugin entry point." The comment did not prevent
PR #352 from exporting a different helper, because comments in a sibling
file are invisible during entry-file edits.

Three coordinated changes:

  1. Move applyBootstrapContent, findBootstrapMarkerBlock, and the
     marker constants into src/lib/bootstrap.ts. The entry point now
     imports the helper. `Object.keys(dist/index.js)` is back to
     `['default']`.

  2. Update tests/unit/plugin.test.ts import path to follow the symbol.

  3. Harden the CI Node ESM smoke test to assert default-only exports
     with an explanatory error message. This is the structural gate
     that would have also caught the v2.5.0 regression, and now
     prevents the entire class for future PRs.

Captures the pattern in
docs/solutions/integration-issues/opencode-plugin-named-exports-break-loader-2026-05-11.md
— including cross-references to the v2.5.0 incident, the PR #335 origin
of `plugInOnce`, and the v2.9.2 agent-colors regression that shares the
same "CI green → npm publish → user install breaks" failure mode.
