+++ .github/workflows/main.yaml
@@ -49,7 +49,15 @@ jobs:
         run: |
           node --input-type=module -e "
             const m = await import('./dist/index.js');
+            const exportKeys = Object.keys(m).sort();
+            const expected = ['default'];
+            if (JSON.stringify(exportKeys) !== JSON.stringify(expected)) {
+              throw new Error('Plugin entry must export only \`default\`. Got: ' + JSON.stringify(exportKeys) + '. Extra named exports break OpenCode plugin loading (the loader treats them as additional plugin factories and crashes when invoking them with no input). Move helpers to src/lib/.');
+            }
             const pluginFactory = m.default;
+            if (typeof pluginFactory !== 'function') {
+              throw new Error('Default export must be a function. Got: ' + typeof pluginFactory);
+            }
             const plugin = await pluginFactory({
               client: { app: { log: async () => {} } },
               directory: process.cwd()
